+++ data/deduplicator/truncate_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/tidepool-org/platform/data"
 	"github.com/tidepool-org/platform/data/deduplicator"
 	testDataStoreDEPRECATED "github.com/tidepool-org/platform/data/storeDEPRECATED/test"
-	testData "github.com/tidepool-org/platform/data/test"
+	dataTest "github.com/tidepool-org/platform/data/test"
 	"github.com/tidepool-org/platform/data/types/upload"
 	"github.com/tidepool-org/platform/log"
 	"github.com/tidepool-org/platform/log/null"
@@ -38,13 +38,13 @@ var _ = Describe("Truncate", func() {
 			testFactory, err = deduplicator.NewTruncateFactory()
 			Expect(err).ToNot(HaveOccurred())
 			Expect(testFactory).ToNot(BeNil())
-			testUploadID = data.NewSetID()
+			testUploadID = dataTest.RandomSetID()
 			testUserID = userTest.RandomID()
 			testDataSet = upload.New()
 			Expect(testDataSet).ToNot(BeNil())
 			testDataSet.UploadID = &testUploadID
 			testDataSet.UserID = &testUserID
-			testDataSet.DeviceID = pointer.FromString(testData.NewDeviceID())
+			testDataSet.DeviceID = pointer.FromString(dataTest.NewDeviceID())
 			testDataSet.DeviceManufacturers = pointer.FromStringArray([]string{"Animas"})
 		})