+++ data/test/data_set.go
@@ -1,0 +1,14 @@
+package test
+
+import (
+	"github.com/tidepool-org/platform/data"
+	"github.com/tidepool-org/platform/test"
+)
+
+func RandomSetID() string {
+	return data.NewSetID()
+}
+
+func RandomSetIDs() []string {
+	return test.RandomStringArrayFromRangeAndGeneratorWithoutDuplicates(1, 3, RandomSetID)
+}